% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET.R
\name{ActualET.Gr4j}
\alias{ActualET.Gr4j}
\title{Actual evapotranspiration in GR4J}
\usage{
\method{ActualET}{Gr4j}(InData, Param, ...)
}
\arguments{
\item{InData}{9-list of:
\itemize{
\item ReferenceEvap,
\item PrecipitationHoch,
\item MoistureVolume,
}}

\item{Param}{1-list of:
\itemize{
\item Gr4j_X1
}}

\item{...}{other Parmeters}
}
\value{
Actual evapotranspiration
}
\description{
Actual evapotranspiration in GR4J
}
\references{
https://webgr.inrae.fr/en/models/daily-hydrological-model-gr4j/description-of-the-gr4j-model/

Perrin, C., Michel, C. and Andréassian, V., 2003. Improvement of a parsimonious model for streamflow simulation. Journal of Hydrology, 279 : 275-289, DOI: 10.1016/S0022-1694(03)00225-7
}
