% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET.R
\name{ActualET}
\alias{ActualET}
\title{Actual evapotranspiration}
\usage{
ActualET(InData, ...)
}
\arguments{
\item{InData}{indata list, use SNOWIntercept(runMode = "VIEW") view the variables and theirs structures}

\item{...}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,}
}
\value{
Actual evapotranspiration
}
\description{
Actual evapotranspiration
}
