% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET.R
\name{ActualET.Vic}
\alias{ActualET.Vic}
\title{Actual evapotranspiration in VIC}
\usage{
\method{ActualET}{Vic}(InData, Param, ...)
}
\arguments{
\item{InData}{9-list of:
\itemize{
\item ReferenceEvap,
\item PrecipitationHoch,
\item MoistureVolume,
\item CapacityMax,
\item Volume1,
\item CapacityMax1,
\item AerodynamicResistance,
\item ArchitecturalResistance,
\item StomatalResistance,
}}

\item{Param}{1-list of:
\itemize{
\item paramSoilMoistureCapacityB
}}

\item{...}{other Parmeters}
}
\value{
Actual evapotranspiration
}
\description{
Actual evapotranspiration in VIC
}
\references{
Liang X, Lettenmaier D P. A simple hydrologically based model of land surface water and energy fluxes for general circulation models[J]. Journal of Geophysical Research, 1994(99):415-428.
}
