% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnergyBalance.R
\name{advected_sensible_heat}
\alias{advected_sensible_heat}
\title{advected_sensible_heat}
\usage{
advected_sensible_heat(SnowCoverFract, AirDens, Tair, TGrnd, Ra)
}
\arguments{
\item{SnowCoverFract}{Snow Cover Fraction}

\item{AirDens}{atmos density}

\item{Tair}{air tempratur}

\item{TGrnd}{grund sufface tempratur}

\item{Ra}{AerodynaResist snow surface}
}
\value{
advected sensible heat
}
\description{
Compute the sensible heat flux advected from bare soil patches to
snow covered patches.
}
