% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aerodynamics.R
\name{Aerodyna.SNOWVIC}
\alias{Aerodyna.SNOWVIC}
\title{SNOWAerodyna}
\usage{
Aerodyna.SNOWVIC(InData, Param)
}
\arguments{
\item{InData}{indata list, use SNOWAerodyna(runMode = "VIEW") view the variables and theirs structures}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,
the other parameters depednd on the actuell model, eg. TimeStepSec, gridN.
and use SNOWAerodyna(runMode = "VIEW") view the structure}
}
\value{
use Aerodyna.SNOWVIC(runMode = "VIEW") view the outputs and theirs structure
}
\description{
Calculate the aerodynamic resistance for each vegetation layer,
and the wind 2m above the layer boundary.  In case of an overstory,
also calculate the wind in the overstory. The values are
normalized based on a reference height wind speed, WindUref, of 1 m/s.
To get wind speeds and aerodynamic resistances for other values of
WindUref, you need to multiply the here calculated wind speeds by WindUref
and divide the here calculated aerodynamic resistances by WindUref.
}
