% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physic.R
\name{air_density}
\alias{air_density}
\title{air_density}
\usage{
air_density(t, p)
}
\arguments{
\item{t}{temperature}

\item{p}{pressure}
}
\value{
rho surface pressure
}
\description{
convert surface pressure (Pa) to density (kg/m3) based on
         pressure (p), vapor pressure (vp), and temperature
}
