% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASEFLOW.R
\name{BASEFLOW.ARNO}
\alias{BASEFLOW.ARNO}
\title{baseflow}
\usage{
\method{BASEFLOW}{ARNO}(InData, Param, ...)
}
\arguments{
\item{InData}{2-list of:
\itemize{
\item SoilMoistureVolume
\item SoilMoistureVolumeMax
}}

\item{Param}{4-list of:
\itemize{
\item paDrainageLossMax
\item paDrainageLossMin
\item paExponentARNOBase
\item paSoilMoistureVolumeARNOBaseThresholdRadio
}}

\item{...}{other Parmeters}
}
\value{
baseflow
}
\description{
baseflow
}
\references{
Arnold J G, Srinivasan R, Muttiah R S et al. Large area hydrologic modeling and assessmentpart I: Model Develoment [J]. Journal of the American Water Resources Association, 1998(34):73-89.
}
