% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASEFLOW.R
\name{BASEFLOW.Arnold}
\alias{BASEFLOW.Arnold}
\title{baseflow}
\usage{
\method{BASEFLOW}{Arnold}(InData, Param, runMode = "RUN", viewGN = 3, ...)
}
\arguments{
\item{InData}{2-list of:
\itemize{
\item SoilMoistureVolume
\item SoilMoistureVolumeMax
}}

\item{Param}{4-list of:
\itemize{
\item paDrainageLossMax
\item paDrainageLossMin
\item paExponentARNOBase
\item paSoilMoistureVolumeARNOBaseThresholdRadio
}}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}

\item{...}{other Parmeters}
}
\value{
baseflow
}
\description{
baseflow
}
\references{
Arnold J G, Srinivasan R, Muttiah R S et al. Large area hydrologic modeling and assessmentpart I: Model Develoment [J]. Journal of the American Water Resources Association, 1998(34):73-89.
}
