% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASEFLOW.R
\name{BASEFLOW}
\alias{BASEFLOW}
\title{caculate Base flow}
\usage{
BASEFLOW(InData, ...)
}
\arguments{
\item{InData}{indata list, use SNOWIntercept(runMode = "VIEW") view the variables and theirs structures}

\item{...}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,}
}
\value{
BASEFLOW
}
\description{
caculate Base flow
}
