% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physic.R
\name{calc_latent_heat_of_sublimation}
\alias{calc_latent_heat_of_sublimation}
\title{calc_latent_heat_of_sublimation}
\usage{
calc_latent_heat_of_sublimation(temp)
}
\arguments{
\item{temp}{temprature}
}
\value{
latent_heat_of_sublimation
}
\description{
Compute the latent heat of sublimation.
}
\references{
Eq. 3.19, Bras 1990
}
