% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physic.R
\name{calc_latent_heat_of_vaporization}
\alias{calc_latent_heat_of_vaporization}
\title{calc_latent_heat_of_vaporization}
\usage{
calc_latent_heat_of_vaporization(temp)
}
\arguments{
\item{temp}{temprature}
}
\value{
latent_heat_of_vaporization
}
\description{
This subroutine computes the latent heat of vaporization.
}
\references{
Eq. 4.2.1 in Handbook of Hydrology.
}
