% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physic.R
\name{calc_outgoing_longwave}
\alias{calc_outgoing_longwave}
\title{calc_outgoing_longwave}
\usage{
calc_outgoing_longwave(temp, emiss)
}
\arguments{
\item{temp}{temprature}

\item{emiss}{Snow emissivity}
}
\value{
outgoing_longwave
}
\description{
Compute outgoing longwave using the Stefan-Boltzman Law.
}
