% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation.R
\name{calc_rainonly}
\alias{calc_rainonly}
\title{calc_rainonly}
\usage{
calc_rainonly(air_temp, prec, MAX_SNOW_TEMP, MIN_RAIN_TEMP)
}
\arguments{
\item{air_temp}{air_temp}

\item{prec}{precptation}

\item{MAX_SNOW_TEMP}{MAX_SNOW_TEMP}

\item{MIN_RAIN_TEMP}{MIN_RAIN_TEMP}
}
\value{
rainfall only
}
\description{
Determines from the air temperature what fraction of incoming
precipitation is frozen and unfrozen (snow and rain).
}
