% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physic.R
\name{calc_saturated_vapor_pressure}
\alias{calc_saturated_vapor_pressure}
\title{calc_saturated_vapor_pressure}
\usage{
calc_saturated_vapor_pressure(temp, Param)
}
\arguments{
\item{temp}{air temprature}

\item{Param}{Param list}
}
\value{
saturated_vapor_pressure
}
\description{
This routine computes the saturated vapor pressure
}
\references{
Handbook of Hydrology eqn 4.2.2.
}
