% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physic.R
\name{calc_scale_height}
\alias{calc_scale_height}
\title{calc_scale_height}
\usage{
calc_scale_height(tair, elevation, Param)
}
\arguments{
\item{tair}{air temprature}

\item{elevation}{elevation}

\item{Param}{params list}
}
\value{
scale_height
}
\description{
Calculate scale height based on average temperature in the column
}
