% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnergyBalance.R, R/physic.R
\name{calc_sensible_heat}
\alias{calc_sensible_heat}
\title{calc_sensible_heat}
\usage{
calc_sensible_heat(atmos_density, t1, t0, Ra)

calc_sensible_heat(atmos_density, t1, t0, Ra)
}
\arguments{
\item{atmos_density}{atmos density}

\item{t1}{tempratur}

\item{t0}{tempratur}

\item{Ra}{AerodynaResist}
}
\value{
sensible_heat

scale_height
}
\description{
Compute the sensible heat flux.

Compute the sensible heat flux.
}
