% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOWOther.R
\name{calc_snow_coverage}
\alias{calc_snow_coverage}
\title{calc_snow_coverage}
\usage{
calc_snow_coverage(
  snowfall,
  melt,
  store_swq,
  swq,
  old_swq,
  store_coverage,
  old_coverage,
  depth,
  old_depth,
  max_snow_distrib_slope
)
}
\arguments{
\item{snowfall}{snowfall}

\item{melt}{melt}

\item{store_swq}{store_swq}

\item{swq}{swq}

\item{old_swq}{old_swq}

\item{store_coverage}{store_coverage}

\item{old_coverage}{old_coverage}

\item{depth}{snow depth}

\item{old_depth}{old_depth}

\item{max_snow_distrib_slope}{max_snow_distrib_slope}
}
\value{
snow coverage
}
\description{
This routine computes the current fraction of the vegetation band that is
covered with snow.  The snow distribution is assumed to be uniform with a
slope based on the value of max_snow_distrib_slope.  The
original value was based on field observations from the University of Minnesota's
Rosemount Agricultural Experiment station (see dissertation by Keith
Cherkauer, 2001).
}
