% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOWBlowing.R
\name{CalcSubFlux}
\alias{CalcSubFlux}
\title{CalcSubFlux}
\usage{
CalcSubFlux(
  EactAir,
  es,
  Zrh,
  AirDens,
  utshear,
  ushear,
  fe,
  U10,
  Zo_salt,
  FT,
  Param,
  Options
)
}
\arguments{
\item{EactAir}{not klar}

\item{es}{not klar}

\item{Zrh}{not klar}

\item{AirDens}{not klar}

\item{utshear}{not klar}

\item{ushear}{not klar}

\item{fe}{not klar}

\item{U10}{not klar}

\item{Zo_salt}{not klar}

\item{FT}{not klar}

\item{Param}{parameters list}

\item{Options}{Options list}
}
\value{
sublimation flux
}
\description{
Calculate the sublimation flux.
}
