% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnergyBalance.R
\name{calcSurfSnowRestTerm}
\alias{calcSurfSnowRestTerm}
\title{calcSurfSnowRestTerm}
\usage{
calcSurfSnowRestTerm(TSurf, InData, Param)
}
\arguments{
\item{TSurf}{tempratur of snowfurface}

\item{InData}{list of Input data}

\item{Param}{Paramaters and inputdata}
}
\value{
a out list
}
\description{
Calculate the surface energy balance for the snow pack.
Virtual function of SnowPackEnergyBalance
}
