% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATE.R
\name{CALIBRATE}
\alias{CALIBRATE}
\title{Get an efficiency coefficient from the parameter group}
\usage{
CALIBRATE(ParamterCalibrate, fctEvaluate = NSE, InList, Observe, warmUPN = 730)
}
\arguments{
\item{ParamterCalibrate}{parameter group}

\item{fctEvaluate}{= NSE(Nash effect) The function to calculate the efficiency coefficient. It is recommended to be found in the "hydroGOF" R-package or written by yourself.}

\item{InList}{InList for hydro Model}

\item{Observe}{the observe data}

\item{warmUPN}{= 730 the warm up period, they will not be evaluated}
}
\value{
coefficient Wert bei selecte coefficient function and Model
}
\description{
Get an efficiency coefficient from the parameter group
}
