% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnergyBalance.R
\name{CanopyEnergyBalance}
\alias{CanopyEnergyBalance}
\title{CanopyEnergyBalance}
\usage{
CanopyEnergyBalance(InData, Param)
}
\arguments{
\item{InData}{list of Input data}

\item{Param}{Paramaters and inputdata}
}
\value{
a out list
}
\description{
This routine iterates to determine the temperature of the canopy, and solve
the resulting fluxes between the canopy and the atmosphere and the canopy
and the ground.
}
