% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUTE.R
\name{CONFLUENCE.G2RES}
\alias{CONFLUENCE.G2RES}
\title{caculate route with UH}
\usage{
\method{CONFLUENCE}{G2RES}(RunoffFlowList, UHAll, TypeGridID, TransAll, ...)
}
\arguments{
\item{RunoffFlowList}{list of Inputdata(Runoff)}

\item{UHAll}{all of the UH}

\item{TypeGridID}{all of Grid type in geological matrix}

\item{TransAll}{all of translate matrix}

\item{...}{other Paramater and inputdata}
}
\value{
StationFlow Q
}
\description{
caculate route with UH
}
