% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUTE.R
\name{CONFLUENCE}
\alias{CONFLUENCE}
\title{caculate route with UH}
\usage{
CONFLUENCE(RunoffFlowList, UHAll, ...)
}
\arguments{
\item{RunoffFlowList}{list of Inputdata(Runoff)}

\item{UHAll}{all of the UH}

\item{...}{other Paramater and inputdata}
}
\value{
StationFlow Q
}
\description{
caculate route with UH
}
