% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_BASEFLOW.ARNO}
\alias{Data_BASEFLOW.ARNO}
\title{Data_BASEFLOW.ARNO}
\format{
list of function data(InData, Param, OutData)
}
\usage{
Data_BASEFLOW.ARNO
}
\description{
Data_BASEFLOW.ARNO
}
\keyword{datasets}
