% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_GROUNDWATER.Vic}
\alias{Data_GROUNDWATER.Vic}
\title{Data_GROUNDWATER.Vic}
\format{
list of function data(InData, Param, OutData)
}
\usage{
Data_GROUNDWATER.Vic
}
\description{
Data_GROUNDWATER.Vic
}
\keyword{datasets}
