% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_Infiltration.OIER}
\alias{Data_Infiltration.OIER}
\title{Data_Infiltration.OIER.}
\format{
list of function data(InData, Param, OutData)
}
\usage{
Data_Infiltration.OIER
}
\description{
Data_Infiltration.OIER.
}
\keyword{datasets}
