% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_InfiltratRat.GreenAmpt}
\alias{Data_InfiltratRat.GreenAmpt}
\title{Data_InfiltratRat.GreenAmpt.}
\format{
list of function data(InData, Param, OutData)
}
\usage{
Data_InfiltratRat.GreenAmpt
}
\description{
Data_InfiltratRat.GreenAmpt.
}
\keyword{datasets}
