% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_INTERCEPTION.Gash}
\alias{Data_INTERCEPTION.Gash}
\title{Data_INTERCEPTION.Gash}
\format{
list of function data(InData, Param, OutData)
}
\usage{
Data_INTERCEPTION.Gash
}
\description{
Data_INTERCEPTION.Gash
}
\keyword{datasets}
