% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_ReferenceET.Hargreaves}
\alias{Data_ReferenceET.Hargreaves}
\title{Data_ReferenceET.Hargreaves.}
\format{
list of function data(InData, Param, OutData)
}
\usage{
Data_ReferenceET.Hargreaves
}
\description{
Data_ReferenceET.Hargreaves.
}
\keyword{datasets}
