% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_ROUTE.Gr4j}
\alias{Data_ROUTE.Gr4j}
\title{Data_ROUTE.Gr4j}
\format{
list of function data(InData, Param, OutData)
}
\usage{
Data_ROUTE.Gr4j
}
\description{
Data_ROUTE.Gr4j
}
\keyword{datasets}
