% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_RUNOFF.OIER}
\alias{Data_RUNOFF.OIER}
\title{Data_RUNOFF.OIER}
\format{
list of function data(InData, Param, OutData)
}
\usage{
Data_RUNOFF.OIER
}
\description{
Data_RUNOFF.OIER
}
\keyword{datasets}
