% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_RUNOFF.SER}
\alias{Data_RUNOFF.SER}
\title{Data_RUNOFF.SER.}
\format{
list of function data(InData, Param, OutData)
}
\usage{
Data_RUNOFF.SER
}
\description{
Data_RUNOFF.SER.
}
\keyword{datasets}
