% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET.R
\name{ET}
\alias{ET}
\title{Actual evapotranspiration}
\usage{
ET(InList, ...)
}
\arguments{
\item{InList}{list of inputData}

\item{...}{other Paramater and inputdata}
}
\value{
Actual evapotranspiration
}
\description{
Actual evapotranspiration
}
