% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET.R
\name{ET.VIC}
\alias{ET.VIC}
\title{Actual evapotranspiration in VIC}
\usage{
\method{ET}{VIC}(InList, PaList, ...)
}
\arguments{
\item{InList}{9-list of:
\itemize{
\item ReferenceEvap,
\item PrecipitationHoch,
\item MoistureVolume,
\item CapacityMax,
\item Volume1,
\item CapacityMax1,
\item AerodynamicResistance,
\item ArchitecturalResistance,
\item StomatalResistance,
}}

\item{PaList}{1-list of:
\itemize{
\item paramSoilMoistureCapacityB
}}

\item{...}{other Paramater and inputdata}
}
\value{
Actual evapotranspiration
}
\description{
Actual evapotranspiration in VIC
}
\references{
Liang X, Lettenmaier D P. A simple hydrologically based model of land surface water and energy fluxes for general circulation models[J]. Journal of Geophysical Research, 1994(99):415-428.
}
