% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALIBRATE.R
\name{EVALUATE}
\alias{EVALUATE}
\title{Get an efficiency coefficient from the parameter group}
\usage{
EVALUATE(
  Paramter_Calibrate,
  fct_Run_Model,
  TimeVariData,
  TimeInvariData,
  Paramter_fest,
  Observe,
  warmUPN = 300,
  fct_Evaluate = NSE
)
}
\arguments{
\item{Paramter_Calibrate}{parameter for the parameter that need to calibrate}

\item{fct_Run_Model}{function for Run_Model}

\item{TimeVariData}{Time Vari Data}

\item{TimeInvariData}{Time invariabel Data}

\item{Paramter_fest}{parameter for the parameter that don't need to calibrate}

\item{Observe}{the observe data}

\item{warmUPN}{= 730 the warm up period, they will not be evaluated}

\item{fct_Evaluate}{= NSE(Nash effect) The function to calculate the efficiency coefficient. It is recommended to be found in the "hydroGOF" R-package or written by yourself.}
}
\value{
coefficient Wert bei selecte coefficient function and Model
}
\description{
Get an efficiency coefficient from the parameter group
}
