% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET.R
\name{fctAerodynamicResistance}
\alias{fctAerodynamicResistance}
\title{AerodynamicResistance paramter caculate}
\usage{
fctAerodynamicResistance(
  VegetationDisplacementHeight,
  VegetationRoughLength,
  WindSpeed
)
}
\arguments{
\item{VegetationDisplacementHeight}{Vegetation Displacement Height}

\item{VegetationRoughLength}{Vegetation Rough Length}

\item{WindSpeed}{Wind Speed}
}
\value{
AerodynamicResistance
}
\description{
AerodynamicResistance paramter caculate
}
