% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUTE.R
\name{fctCutGridFlow}
\alias{fctCutGridFlow}
\title{cut GridFlow from AllGridFlow to AimGridFlow and OtherGridFlow, from 1 cut to 2. Get list with $OtherGridFlow and $AimGridFlow}
\usage{
fctCutGridFlow(OriginalGridFlow, GridID, AimID)
}
\arguments{
\item{OriginalGridFlow}{flow of origal Grid}

\item{GridID}{origal Grid ID}

\item{AimID}{aim Grid ID}
}
\value{
2-list of:
\itemize{
\item FlowOtherGrid
\item FlowAimGrid
}
}
\description{
cut GridFlow from AllGridFlow to AimGridFlow and OtherGridFlow, from 1 cut to 2. Get list with $OtherGridFlow and $AimGridFlow
}
