% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUTE.R
\name{fctG2AimGAll}
\alias{fctG2AimGAll}
\title{caculate all Grid to next father Grid}
\usage{
fctG2AimGAll(TypeGridID, GridID, FlowDirection, GridDEM)
}
\arguments{
\item{TypeGridID}{all of Grid type in geological matrix}

\item{GridID}{ID in geological Grid}

\item{FlowDirection}{Flow Direction in geological Grid}

\item{GridDEM}{DEM in geological Grid}
}
\value{
list of TypeGridN - 1
}
\description{
caculate all Grid to next father Grid
}
