% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUTE.R
\name{fctGrid2AimGridFind}
\alias{fctGrid2AimGridFind}
\title{find every Grid which Grid will go (confluence way) prepare for UH make}
\usage{
fctGrid2AimGridFind(l, m, AimGridID, GridID, FlowDirection)
}
\arguments{
\item{l}{num}

\item{m}{num}

\item{AimGridID}{Aim}

\item{GridID}{orignal}

\item{FlowDirection}{FlowDirection}
}
\value{
vector of fluss
}
\description{
find every Grid which Grid will go (confluence way) prepare for UH make
}
