% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUTE.R
\name{fctGTransMatAll}
\alias{fctGTransMatAll}
\title{caculate all Grid to next father Grid}
\usage{
fctGTransMatAll(G2AimGList, TypeGridID)
}
\arguments{
\item{G2AimGList}{all of matrix of origal Grid to aim Grid}

\item{TypeGridID}{all of Grid type in geological matrix}
}
\value{
list of TypeGridN - 1
}
\description{
caculate all Grid to next father Grid
}
