% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GROUNDWATER.R
\name{fctHydraulicConductivity}
\alias{fctHydraulicConductivity}
\title{soil paramter caculate}
\usage{
fctHydraulicConductivity(
  VolumetricSoilMoistureContent,
  SoilPorosity,
  SaturatedHydraulicConductivity,
  paramClappHornbergerB
)
}
\arguments{
\item{VolumetricSoilMoistureContent}{Volumetric Soil Moisture Content}

\item{SoilPorosity}{Soil Porosity}

\item{SaturatedHydraulicConductivity}{Saturated Hydraulic Conductivity}

\item{paramClappHornbergerB}{ClappHornberger B}
}
\value{
HydraulicConductivity
}
\description{
soil paramter caculate
}
