% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GROUNDWATER.R
\name{fctHydraulicDiffusivity}
\alias{fctHydraulicDiffusivity}
\title{soil paramter caculate}
\usage{
fctHydraulicDiffusivity(
  VolumetricSoilMoistureContent,
  SoilPorosity,
  SaturatedSoilSuctionHead,
  SaturatedHydraulicConductivity,
  paramClappHornbergerB
)
}
\arguments{
\item{VolumetricSoilMoistureContent}{Soil Moisture Content}

\item{SoilPorosity}{SoilPorosity}

\item{SaturatedSoilSuctionHead}{Saturated Soil SuctionHead}

\item{SaturatedHydraulicConductivity}{Saturated Hydraulic Conductivity}

\item{paramClappHornbergerB}{param Clapp Hornberger B}
}
\value{
HydraulicDiffusivity
}
\description{
soil paramter caculate
}
