% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{fctInfiltration.GreenAmpt}
\alias{fctInfiltration.GreenAmpt}
\title{Infiltration with GreenAmpt methond}
\usage{
\method{fctInfiltration}{GreenAmpt}(InList, ...)
}
\arguments{
\item{InList}{5-list of:
\itemize{
\item HydraulicConductivity
\item WettingFrontSoilSuction
\item SoilMoistureContent
\item EffectivePorosity
\item SoilMoistureVolume
}}

\item{...}{other Paramater and inputdata}
}
\value{
Infiltration
}
\description{
Infiltration with GreenAmpt methond
}
\references{
Tarboton D G. Rainfall-runoff processes[M]. Utah State University, 2003.
}
