% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{fctInfiltration}
\alias{fctInfiltration}
\title{Infiltration caculate}
\usage{
fctInfiltration(InList, ...)
}
\arguments{
\item{InList}{list of Input data}

\item{...}{other Paramater and inputdata}
}
\value{
Infiltration
}
\description{
Infiltration caculate
}
