% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUH.R
\name{fctIUH.Nash}
\alias{fctIUH.Nash}
\title{Instantaneous unit hydrograph in Nash}
\usage{
\method{fctIUH}{Nash}(t, InList)
}
\arguments{
\item{t}{numrica,time}

\item{InList}{list of Input data}
}
\value{
Instantaneous unit hydrograph
}
\description{
Instantaneous unit hydrograph in Nash
}
\references{
Nash J E, Sutcliffe J V. River flow forecasting through conceptual models part I — A discussion of principles[J]. Journal of Hydrology, 1970(10):282-290.
}
