% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUH.R
\name{fctIUH}
\alias{fctIUH}
\title{Instantaneous unit hydrograph}
\usage{
fctIUH(t, InList)
}
\arguments{
\item{t}{numrica,time}

\item{InList}{list of Input data}
}
\value{
Instantaneous unit hydrograph
}
\description{
Instantaneous unit hydrograph
}
