% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUH.R
\name{fctIUH.RuiXF}
\alias{fctIUH.RuiXF}
\title{Instantaneous unit hydrograph in RUI XioaFang}
\usage{
\method{fctIUH}{RuiXF}(t, InList)
}
\arguments{
\item{t}{numrica,time}

\item{InList}{list of Input data}
}
\value{
Instantaneous unit hydrograph
}
\description{
Instantaneous unit hydrograph in RUI XioaFang
}
\references{
芮孝芳, 石朋. 基于地貌扩散和水动力扩散的流域瞬时单位线研究[J]. 水科学进展, 2002(13):439-444.
}
