% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUTE.R
\name{fctMakeGridTranslateMatrix}
\alias{fctMakeGridTranslateMatrix}
\title{make the TranslateMatrix, which will mit Acculate from GridList to AimGridList.}
\usage{
fctMakeGridTranslateMatrix(Grid2AimGrid, AimID)
}
\arguments{
\item{Grid2AimGrid}{matrix of origal Grid to aim Grid}

\item{AimID}{aim Grid ID}
}
\value{
TranslateMatrix
}
\description{
make the TranslateMatrix, which will mit Acculate from GridList to AimGridList.
}
