% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUH.R
\name{fctMakeUH}
\alias{fctMakeUH}
\title{make Discrete unit hydrograph durch Instantaneous unit hydrograph}
\usage{
fctMakeUH(G2AimGPa, UHPeriodN)
}
\arguments{
\item{G2AimGPa}{vector of Paramater for IUH}

\item{UHPeriodN}{howmany Discrete have}
}
\value{
matrix, Discrete unit hydrograph
}
\description{
make Discrete unit hydrograph durch Instantaneous unit hydrograph
}
