% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUH.R
\name{fctMakeUHParam}
\alias{fctMakeUHParam}
\title{make Param list for UH}
\usage{
fctMakeUHParam(PaInList, methond)
}
\arguments{
\item{PaInList}{list of parameter for every IUH}

\item{methond}{IUH methond name}
}
\value{
Parameter for UH mit class name
}
\description{
make Param list for UH
}
