% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUH.R
\name{fctMakeUHParamAll}
\alias{fctMakeUHParamAll}
\title{make all Param list for UH all}
\usage{
fctMakeUHParamAll(PaInList, methondList)
}
\arguments{
\item{PaInList}{list of parameter for every IUH}

\item{methondList}{list of IUH methond name}
}
\value{
list of Parameter for all UH mit class name
}
\description{
make all Param list for UH all
}
