% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{fctMoistureCapacity}
\alias{fctMoistureCapacity}
\title{soil paramter caculate}
\usage{
fctMoistureCapacity(
  SoilMoistureVolume,
  SoilMoistureCapacityMax,
  paSoilMoistureCapacityB
)
}
\arguments{
\item{SoilMoistureVolume}{Soil Moisture Volume}

\item{SoilMoistureCapacityMax}{Soil Moisture Capacity Max}

\item{paSoilMoistureCapacityB}{param Soil Moisture Capacity B}
}
\value{
SoilMoistureCapacity
}
\description{
soil paramter caculate
}
