% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{fctSaturatedArea}
\alias{fctSaturatedArea}
\title{soil paramter caculate}
\usage{
fctSaturatedArea(
  SoilMoistureCapacity,
  SoilMoistureCapacityMax,
  paSoilMoistureCapacityB
)
}
\arguments{
\item{SoilMoistureCapacity}{Soil Moisture Capacity}

\item{SoilMoistureCapacityMax}{Soil Moisture Capacity Max}

\item{paSoilMoistureCapacityB}{param Soil Moisture Capacity B}
}
\value{
SaturatedArea
}
\description{
soil paramter caculate
}
