% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{fctSoilInfiltration.OIER}
\alias{fctSoilInfiltration.OIER}
\title{Infiltration rate in Over Infiltration Excess Runoff}
\usage{
\method{fctSoilInfiltration}{OIER}(InList, PaList, ...)
}
\arguments{
\item{InList}{2-list of:
\itemize{
\item PrecipitationHoch
\item InfiltrationRateMax
}}

\item{PaList}{1-list of:
\itemize{
\item paInfiltrationRateB
}}

\item{...}{other Paramater and inputdata}
}
\value{
Infiltration rate in Over Infil tration Excess Runoff
}
\description{
Infiltration rate in Over Infiltration Excess Runoff
}
