% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{fctSoilInfiltration}
\alias{fctSoilInfiltration}
\title{SoilInfiltration rate}
\usage{
fctSoilInfiltration(InList, ...)
}
\arguments{
\item{InList}{list of Input data}

\item{...}{other Paramater and inputdata}
}
\value{
SoilInfiltration
}
\description{
SoilInfiltration rate
}
