% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{fctSoilInfiltration.SER}
\alias{fctSoilInfiltration.SER}
\title{SoilInfiltration rate in Saturation Excess Runoff modell}
\usage{
\method{fctSoilInfiltration}{SER}(InList, PaList, ...)
}
\arguments{
\item{InList}{3-list of:
\itemize{
\item PrecipitationHoch
\item SoilMoistureCapacityMax
\item SoilMoistureCapacity
}}

\item{PaList}{1-list of:
\itemize{
\item paSoilMoistureCapacityB
}}

\item{...}{other Paramater and inputdata}
}
\value{
SoilInfiltration in SaturationExcessRunoff modell
}
\description{
SoilInfiltration rate in Saturation Excess Runoff modell
}
