% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUH.R
\name{fctUHALLMake}
\alias{fctUHALLMake}
\title{make all Discrete unit hydrograph durch Instantaneous unit hydrograph}
\usage{
fctUHALLMake(G2AimGList, UHPeriodN = 180)
}
\arguments{
\item{G2AimGList}{list of all vector for IUH}

\item{UHPeriodN}{howmany Discrete have}
}
\value{
matrix, Discrete unit hydrograph
}
\description{
make all Discrete unit hydrograph durch Instantaneous unit hydrograph
}
