% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUTE.R
\name{fctUHConfluence}
\alias{fctUHConfluence}
\title{caculate the Runoff mit selbst UH, but not have to the AimGrid, just know how many Confluence have to its AimGrid}
\usage{
fctUHConfluence(Flow, UH)
}
\arguments{
\item{Flow}{flow of origal Grid}

\item{UH}{Discrete unit hydrograph}
}
\value{
FlowResult
}
\description{
caculate the Runoff mit selbst UH, but not have to the AimGrid, just know how many Confluence have to its AimGrid
}
