% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{fctVICRateFind}
\alias{fctVICRateFind}
\title{find the rate of SER and OIER}
\usage{
fctVICRateFind(
  rate,
  PrecipitationHoch,
  SoilMoistureCapacityMax,
  SoilMoistureCapacity,
  InfiltrationRateMax,
  Param
)
}
\arguments{
\item{rate}{rate procent}

\item{PrecipitationHoch}{Precipitation Hoch, mm}

\item{SoilMoistureCapacityMax}{Soil Moisture Capacity Max}

\item{SoilMoistureCapacity}{Soil Moisture Capacity}

\item{InfiltrationRateMax}{Infiltration Rate Max}

\item{Param}{2-list of:
\itemize{
\item paSoilMoistureCapacityB
\item paInfiltrationRateB
}}
}
\value{
rate procent
}
\description{
find the rate of SER and OIER
}
