% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metro_funktions.R
\name{get_Dew_point}
\alias{get_Dew_point}
\title{dew point caculate}
\usage{
get_Dew_point(actual_vp)
}
\arguments{
\item{actual_vp}{actual vapor pressure}
}
\value{
dew point temperature
}
\description{
dew point caculate
}
