% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTools.R
\name{get_ModelData}
\alias{get_ModelData}
\title{find the data structure of a model, that new created}
\usage{
get_ModelData(Modules)
}
\arguments{
\item{Modules}{a list, names are the module name and the values are the process methond}
}
\value{
the data structure of the model's data
\itemize{
\item InputData, the data that from outside of the model, must prepared
\item In_InitialData, the data that must judge, wether they from the outside or the state data
\item InData, all InData of all modules
\item OutData, all the OutData of all modules 
\item Param, all parameters that the model needed
}
}
\description{
find the data structure of a model, that new created
}
