% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTools.R
\name{get_ModuleData}
\alias{get_ModuleData}
\title{find the data structure of a module or model}
\usage{
get_ModuleData(funcName, data = "All")
}
\arguments{
\item{funcName}{function(model or module complete name), in character e.g. "Model.CVIC"}

\item{data}{"All" is default, or:"InData", "Param", "OutData",}
}
\value{
the data structure of the function's data
}
\description{
find the data structure of a module or model
}
