% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOWBlowing.R
\name{get_prob}
\alias{get_prob}
\title{get_prob}
\usage{
get_prob(Tair, Age, SurfaceLiquidWater, U10)
}
\arguments{
\item{Tair}{Temperatur of air}

\item{Age}{age of snow in hour}

\item{SurfaceLiquidWater}{SurfaceLiquidWater}

\item{U10}{not klar}
}
\value{
parameters for probability of blowing snow occurence
}
\description{
Calculate parameters for probability of blowing snow occurence.
}
\references{
see Li and Pomeroy 1997
}
