% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metro_funktions.R
\name{get_Saturat_vapor_pressure}
\alias{get_Saturat_vapor_pressure}
\title{Saturat_vapor_pressure caculate}
\usage{
get_Saturat_vapor_pressure(Temperature)
}
\arguments{
\item{Temperature}{Temperature}
}
\value{
Saturat vapor pressure
}
\description{
Saturat_vapor_pressure caculate
}
