% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOWBlowing.R
\name{get_shear}
\alias{get_shear}
\title{get_shear}
\usage{
get_shear(x, Ur, Zr)
}
\arguments{
\item{x}{not klar}

\item{Ur}{not klar}

\item{Zr}{not klar}
}
\value{
the values of all output variables to 0
}
\description{
This routine resets the values of all output variables to 0.
the function modify from VIC original rescouce,
check the rescouce code in https://github.com/UW-Hydro/VIC
}
