% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOWBlowing.R
\name{get_thresh}
\alias{get_thresh}
\title{get_thresh}
\usage{
get_thresh(Tair, SurfaceLiquidWater, Zo_salt)
}
\arguments{
\item{Tair}{Temperatur of air}

\item{SurfaceLiquidWater}{SurfaceLiquidWater}

\item{Zo_salt}{not klar}
}
\value{
threshold shear stress
}
\description{
Calculate threshold shear stress.
}
