% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUH.R
\name{gr4j_UH2}
\alias{gr4j_UH2}
\title{UH2 for RG4J}
\usage{
gr4j_UH2(X4)
}
\arguments{
\item{X4}{parameter X4}
}
\value{
Instantaneous unit hydrograph
}
\description{
UH2 for RG4J
}
\references{
https://webgr.inrae.fr/en/models/daily-hydrological-model-gr4j/description-of-the-gr4j-model/

Perrin, C., Michel, C. and Andréassian, V., 2003. Improvement of a parsimonious model for streamflow simulation. Journal of Hydrology, 279 : 275-289, DOI: 10.1016/S0022-1694(03)00225-7
}
