% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GridList}
\alias{GridList}
\title{GridData 4-list of grid-data used to calculate confluence(route)}
\format{
a list of 4
\describe{
 \item{NOTE}{The field names in these lists must be consistent and will be the only index.}
 \item{TypeGridID}{4-list. Type of each grid point.}
 \item{GridID}{matrix:	ID in grid-data, be sure to mark}{attr(GridID, "Null"), what the Null-Wert in grid-data.}
 \item{FlowDirection}{matrix':	FlowDirection in grid-data, be sure to mark}{attr(GridID, "Null"), what the Null-Wert in grid-data.}
 \item{GridDEM}{matrix':	DEM in grid-data, be sure to mark}{attr(GridID, "Null"), what the Null-Wert in grid-data.}
}
TypeGridID: 4-list
\describe{
 \item{NOTE}{The field names in these lists must be consistent and will be the only index.}
 \item{GridGridID}{all grid ID in 1-array(gridN)}
 \item{RiverGridID}{river grid ID in 1-array, Estuary and HydroStation belong to river}
 \item{EstuaryID}{Estuary ID in 1-array, HydroStation belong to Estuary}
 \item{HydroStationID}{HydroStation ID in 1-array}
}
}
\usage{
GridList
}
\description{
GridData 4-list of grid-data used to calculate confluence(route)
}
\keyword{datasets}
