% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GROUNDWATER.R
\name{GROUNDWATER}
\alias{GROUNDWATER}
\title{GROUNDWATER caculate}
\usage{
GROUNDWATER(InData, ...)
}
\arguments{
\item{InData}{list of Input data}

\item{...}{other Paramater and inputdata}
}
\value{
GROUNDWATER
}
\description{
GROUNDWATER caculate
}
