% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GROUNDWATER.R
\name{GROUNDWATER.VIC}
\alias{GROUNDWATER.VIC}
\title{ground water VIC function}
\usage{
\method{GROUNDWATER}{VIC}(InData, Param, runMode = "RUN", viewGN = 3, ...)
}
\arguments{
\item{InData}{6-list of:
\itemize{
\item Evapotranspiration
\item Interception
\item Infiltration
\item BaseFlow
\item GroundWaterIn
\item GridSoilParame
}}

\item{Param}{1-list of:
\itemize{
\item paClappHornbergerB
}}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}

\item{...}{other Parmeters}
}
\value{
INTERCEPTION
}
\description{
ground water VIC function
}
\references{
Liang X, Xie Z, Huang M. A new parameterization for surface and groundwater interactions and its impact on water budgets with the variable infiltration capacity[J]. Journal of Geochemical Research, 2003(108):1-17.
}
