% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{Infiltration.OIER}
\alias{Infiltration.OIER}
\title{Infiltration rate in Over Infiltration Excess Runoff}
\usage{
\method{Infiltration}{OIER}(InData, Param, ...)
}
\arguments{
\item{InData}{2-list of:
\itemize{
\item PrecipitationHoch
\item InfiltrationRateMax
}}

\item{Param}{1-list of:
\itemize{
\item paInfiltrationRateB
}}

\item{...}{other Parmeters}
}
\value{
Infiltration rate in Over Infil tration Excess Runoff
}
\description{
Infiltration rate in Over Infiltration Excess Runoff
}
