% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{Infiltration.SER}
\alias{Infiltration.SER}
\title{SoilInfiltration rate in Saturation Excess Runoff modell}
\usage{
\method{Infiltration}{SER}(InData, Param, runMode = "RUN", viewGN = 3, ...)
}
\arguments{
\item{InData}{3-list of:
\itemize{
\item PrecipitationHoch
\item SoilMoistureCapacityMax
\item SoilMoistureCapacity
}}

\item{Param}{1-list of:
\itemize{
\item paSoilMoistureCapacityB
}}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}

\item{...}{other Parmeters}
}
\value{
SoilInfiltration in SaturationExcessRunoff modell
}
\description{
SoilInfiltration rate in Saturation Excess Runoff modell
}
