% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{InfiltratRat.GreenAmpt}
\alias{InfiltratRat.GreenAmpt}
\title{Infiltration with GreenAmpt methond}
\usage{
\method{InfiltratRat}{GreenAmpt}(InData, Param, ...)
}
\arguments{
\item{InData}{5-list of:
\itemize{
\item HydraulicConductivity
\item WettingFrontSoilSuction
\item SoilMoistureContent
\item EffectivePorosity
\item SoilMoistureVolume
}}

\item{Param}{Param}

\item{...}{other Parmeters}
}
\value{
Infiltration rat
}
\description{
Infiltration with GreenAmpt methond
}
\references{
Tarboton D G. Rainfall-runoff processes[M]. Utah State University, 2003.
}
