% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{InfiltratRat}
\alias{InfiltratRat}
\title{Infiltration caculate}
\usage{
InfiltratRat(InData, ...)
}
\arguments{
\item{InData}{indata list, use SNOWIntercept(runMode = "VIEW") view the variables and theirs structures}

\item{...}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,}
}
\value{
Infiltration
}
\description{
Infiltration caculate
}
