% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIC.R
\name{InListMake}
\alias{InListMake}
\title{Make the required INList for model}
\usage{
InListMake(ClsName, ...)
}
\arguments{
\item{ClsName}{chr, chr vector, all the methonds of hydrology modul. e.g.ClsNa <- c("VIC", "PenmanMonteith", "GreenAmpt", "Gash", "ARNO", "G2RES")}

\item{...}{other paramters}
}
\value{
InList for model
}
\description{
Make the required list according to the format of the model model input list.
}
