% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIC.R
\name{InListMake.VIC}
\alias{InListMake.VIC}
\title{Make the required INList for VIC}
\usage{
\method{InListMake}{VIC}(
  ClsName,
  infoStarDay,
  infoEndDay,
  MetData,
  GeoData,
  GridData,
  UPMethondList,
  UHPeriodN = 180,
  UHUnitTranslate,
  ...
)
}
\arguments{
\item{ClsName}{chr, chr vector, all the methonds of hydrology modul. e.g.ClsNa <- c("VIC", "PenmanMonteith", "GreenAmpt", "Gash", "ARNO", "G2RES")}

\item{infoStarDay}{chr, the start day, e.g. "1989-1-1"}

\item{infoEndDay}{chr, the end day, e.g. "1993-12-31"}

\item{MetData}{list, num, list of metrol data. For VIC Model should contain at least the following 8 data:
\itemize{
(The field names must be the same as listed in the following list because they will be the only index.)
\item Tmean
\item Tmax
\item Tmin
\item WindSpeed
\item WindH
\item SunHour
\item RelativeHumidity
\item PrecipitationHoch
}
Each data is a 2-array(periodN, gridN).}

\item{GeoData}{list, geological data,  For VIC Model should contain at least the following 4 data:
\itemize{
(The field names must be the same as listed in the following list because they will be the only index.)
\item Evalution: 1-array(gridN). Evalution of all grids.
\item Location: data.frame:  gridN obs. of  3 variables. ID,
\item SoilParam: data.frame':  gridN obs. of  some variables.
\item LanduseParam: data.frame': gridN obs. of  some variables.
}
Location field names, for VIC Model should contain at least the following 2 data:
\itemize{
(The field names must be the same as listed in the following list because they will be the only index.)
\item Latitude
\item Longitude
}
SoilParam field names, for VIC Model should contain at least the following 12 data
\itemize{
(The field names must be the same as listed in the following list because they will be the only index.)
\item T_Porosity_: top soil...
\item T_FieldCapacity_
\item T_WiltingPoint_
\item T_WettingFrontSoilSuctionHead_mm
\item T_SaturatedSoilSuctionHead_mm
\item T_SaturatedHydraulicConductivity_mm_day
\item S_Porosity_: sub soil...
\item S_FieldCapacity_
\item S_WiltingPoint_
\item S_WettingFrontSoilSuctionHead_mm
\item S_SaturatedSoilSuctionHead_mm
\item S_SaturatedHydraulicConductivity_mm_day
}
LanduseParam field names, for VIC Model should contain at least the following 28 data
\itemize{
(The field names must be the same as listed in the following list because they will be the only index.)
\item SL_mm
\item root_depth_mm
\item rarc
\item rmin
\item ROU.JAN - ROU.DEC (*12)
\item DIS.JAN - DIS.DEC (*12)
}}

\item{GridData}{4-list of grid-data used to calculate confluence(route)
\itemize{
(The field names must be the same as listed in the following list because they will be the only index.)
\item TypeGridID: 4-list. Type of each grid point.
\item GridID: matrix:  ID in grid-data
\item FlowDirection: matrix':  FlowDirection in grid-data
\item GridDEM: matrix':  DEM in grid-data
}
TypeGridID: 4-list
\itemize{
(The field names must be the same as listed in the following list because they will be the only index.)
\item GridGridID    : all grid ID in 1-array(gridN)
\item RiverGridID   : river grid ID in 1-array, Estuary and HydroStation belong to river
\item EstuaryID     : Estuary ID in 1-array, HydroStation belong to Estuary
\item HydroStationID: HydroStation ID in 1-array
}}

\item{UPMethondList}{four methonds Name for IUH, e.g. c("Shipeng", "Shipeng", "Shipeng", "Shipeng")}

\item{UHPeriodN}{howmany Period have in Discrete unit hydrograph}

\item{UHUnitTranslate}{Transformation parameters from mm to m^3/s}

\item{...}{other paramters}
}
\value{
InList for VIC
}
\description{
Make the required list according to the format of the model VIC input list.
}
\examples{
ClsNa <- c("PenmanMonteith", "GreenAmpt", "Gash", "ARNO", "G2RES", "VIC")
class(ClsNa) <- tail(ClsNa, 1)
UPMethondList = list("Shipeng", "Shipeng", "Shipeng", "Shipeng")
VICInList <- InListMake(ClsNa,
                            "1989-1-1",
                            "1993-12-31",
                            MetroList,
                            GeoList,
                            GridList,
                            UPMethondList,
                            180,
                            35)
}
