% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERCEPTION.R
\name{INTERCEPTION}
\alias{INTERCEPTION}
\title{INTERCEPTION caculate}
\usage{
INTERCEPTION(InData, ...)
}
\arguments{
\item{InData}{indata list, use SNOWIntercept(runMode = "VIEW") view the variables and theirs structures}

\item{...}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,}
}
\value{
INTERCEPTION
}
\description{
INTERCEPTION caculate
}
