% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnergyBalance.R
\name{latent_heat_from_snow}
\alias{latent_heat_from_snow}
\title{latent_heat_from_snow}
\usage{
latent_heat_from_snow(
  AirDens,
  EactAir,
  Lv,
  Press,
  Ra,
  TMean,
  Vpd,
  BlowingMassFlux,
  Param
)
}
\arguments{
\item{AirDens}{atmos density}

\item{EactAir}{VaporPressure}

\item{Lv}{latent_heat_of_vaporization}

\item{Press}{air Press}

\item{Ra}{AerodynaResist}

\item{TMean}{tempratur}

\item{Vpd}{Vapor Press Defficit}

\item{BlowingMassFlux}{Blowing Mass Flux}

\item{Param}{paramters list}
}
\value{
latent_heat
}
\description{
Calculate the saturated vapor pressure in the snow pack,
(Equation 3.32, Bras 1990)
}
