% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUH.R
\name{makeUHALL}
\alias{makeUHALL}
\title{make all Discrete unit hydrograph durch Instantaneous unit hydrograph}
\usage{
makeUHALL(InData, Param, ...)
}
\arguments{
\item{InData}{indata list, use snow_density(runMode = "VIEW") view the variables and theirs structures}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,}

\item{...}{other Parmeters}
}
\value{
matrix, Discrete unit hydrograph
}
\description{
make all Discrete unit hydrograph durch Instantaneous unit hydrograph
}
