% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOWOther.R
\name{MassRelease}
\alias{MassRelease}
\title{MassRelease}
\usage{
MassRelease(
  InterceptedSnow,
  TempInterceptionStorage,
  ReleasedMass,
  Drip,
  VEG_MIN_INTERCEPTION_STORAGE,
  iterN = 0
)
}
\arguments{
\item{InterceptedSnow}{InterceptedSnow}

\item{TempInterceptionStorage}{TempInterceptionStorage}

\item{ReleasedMass}{ReleasedMass}

\item{Drip}{Drip}

\item{VEG_MIN_INTERCEPTION_STORAGE}{VEG_MIN_INTERCEPTION_STORAGE}

\item{iterN}{not klar}
}
\value{
Release Mass
}
\description{
Calculates mass release of snow from canopy.
}
