% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MetroList}
\alias{MetroList}
\title{MetData list, num, list of metrol data. For VIC Model should contain at least the following 8 data:}
\format{
a list of 8
\describe{
 \item{NOTE}{The field names in these lists must be consistent and will be the only index.}
 \item{Tmean}{mean tempratur}
 \item{Tmax}{max tempratur}
 \item{Tmin}{min tempratur}
 \item{WindSpeed}{Wind Speed}
 \item{WindH}{Wind messure Hight}
 \item{SunHour}{Sun Hour in procent}
 \item{RelativeHumidity}{Relative Humidity}
 \item{PrecipitationHoch}{Precipitation Hoch}
}
    Each data is a 2-array(periodN, gridN).
}
\usage{
MetroList
}
\description{
MetData list, num, list of metrol data. For VIC Model should contain at least the following 8 data:
}
\keyword{datasets}
