% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVIC.R
\name{MODEL.CVIC}
\alias{MODEL.CVIC}
\title{VIC runoff (OverInfiltrationExcessRunoff and SaturationExcessRunoff)}
\usage{
\method{MODEL}{CVIC}(TimeVariData, TimeInvariData, Param, ...)
}
\arguments{
\item{TimeVariData}{Time Vari Data}

\item{TimeInvariData}{Time invariabel Data}

\item{Param}{Param}

\item{...}{other arguments}
}
\value{
Q
}
\description{
VIC runoff (OverInfiltrationExcessRunoff and SaturationExcessRunoff)
}
