% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR4J.R
\name{MODEL.GR4J}
\alias{MODEL.GR4J}
\title{Model GR4J}
\usage{
\method{MODEL}{GR4J}(TimeVariData, TimeInvariData = NULL, Param, ...)
}
\arguments{
\item{TimeVariData}{Time Vari Data}

\item{TimeInvariData}{Time invariabel Data}

\item{Param}{Param}

\item{...}{other arguments}
}
\value{
Q
}
\description{
Model GR4J
}
\references{
https://webgr.inrae.fr/en/models/daily-hydrological-model-gr4j/description-of-the-gr4j-model/

Perrin, C., Michel, C. and Andréassian, V., 2003. Improvement of a parsimonious model for streamflow simulation. Journal of Hydrology, 279 : 275-289, DOI: 10.1016/S0022-1694(03)00225-7
}
