% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVIC.R
\name{MODEL}
\alias{MODEL}
\title{MODEL}
\usage{
MODEL(TimeVariData, TimeInvariData, Param, ...)
}
\arguments{
\item{TimeVariData}{Time Vari Data}

\item{TimeInvariData}{Time invariabel Data}

\item{Param}{Param}

\item{...}{other arguments}
}
\value{
Q
}
\description{
Model
}
