% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIC.R
\name{MODELL}
\alias{MODELL}
\title{modell}
\usage{
MODELL(VorIn, VorPa)
}
\arguments{
\item{VorIn}{input list for model}

\item{VorPa}{pramater list for model}
}
\value{
Q(flow) of hydrostation, m^3/s
}
\description{
modell
}
