% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIC.R
\name{MODELL.VIC}
\alias{MODELL.VIC}
\title{modell VIC}
\usage{
\method{MODELL}{VIC}(VorIn, VorPa)
}
\arguments{
\item{VorIn}{please checke output of InListMake.VIC()}

\item{VorPa}{please checke output of PaListMake.VIC()}
}
\value{
Q(flow) of hydrostation, m^3/s
}
\description{
modell VIC
}
\examples{
ClsNa <- c("PenmanMonteith", "GreenAmpt", "Gash", "ARNO", "G2RES", "VIC")
class(ClsNa) <- tail(ClsNa, 1)
UPMethondList = list("Shipeng", "Shipeng", "Shipeng", "Shipeng")
VICInList <- InListMake(ClsNa,
                            "1989-1-1",
                            "1989-1-6",
                            MetroList,
                            GeoList,
                            GridList,
                            UPMethondList,
                            180,
                            35)
ParamterMax = c(15, 300, 600, 900, 0.9, 2.7, 0.7, 30, 9, 1,1,15,7,7,7,10,10,10,15)
ParamterMin = c(5, 50, 100, 150, 0.1, 1.3, 0.1,9, 0.1, 0.1, 0.1, 3,0.5, 0.5, 0.5 , 0.5,0.5,0.5,0.5)
ParamterModell = 0.5 * (ParamterMax + ParamterMin)
class(ParamterModell) <- ClsNa
VICPaList <- PaListMake(ParamterModell)
Q <- MODELL(VICInList, VICPaList)
}
