% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOWOther.R
\name{new_snow_density}
\alias{new_snow_density}
\title{new_snow_density}
\usage{
new_snow_density(air_temp, Param, runMode = "RUN")
}
\arguments{
\item{air_temp}{air tempteratur}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,
the other parameters depednd on the actuell model, eg. TimeStepSec, gridN.
and use new_snow_density(runMode = "VIEW") view the structure}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}
}
\value{
use new_snow_density(runMode = "VIEW") view the outputs and theirs structure
}
\description{
Estimate the density of new snow
}
