% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIC.R
\name{PaListMake}
\alias{PaListMake}
\title{Make the required PaList for hydro Model}
\usage{
PaListMake(ParamterModell)
}
\arguments{
\item{ParamterModell}{vector of all to be calibrated paramters for hydro Model.}
}
\value{
PaList for hydro Model
}
\description{
Make the required list according to the format of the model paramters list.
}
