% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIC.R
\name{PaListMake.VIC}
\alias{PaListMake.VIC}
\title{Make the required PaList for VIC}
\usage{
\method{PaListMake}{VIC}(ParamterModell)
}
\arguments{
\item{ParamterModell}{vector of all to be calibrated paramters for VIC.
\itemize{
    \item ZoneDepth = ParamterModell[1:4],
    \item paCoefficientFreeThroughfall = ParamterModell[5],
    \item paExponentARNOBase = ParamterModell[6],
    \item paSoilMoistureVolumeARNOBaseThresholdRadio = ParamterModell[7],
    \item paDrainageLossMax = ParamterModell[8],##*## #mm
    \item paDrainageLossMin = ParamterModell[9], ##*## #mm
    \item paSoilMoistureCapacityB = ParamterModell[10],
    \item paInfiltrationRateB = ParamterModell[11],
    \item paClappHornbergerB = ParamterModell[12],
    \item UPPaList = ParamterModell[13:19]
}}
}
\value{
PaList for VIC
}
\description{
Make the required list according to the format of the model VIC paramters list.
}
\examples{
ParamterMax = c(15, 300, 600, 900, 0.9, 2.7, 0.7, 30, 9, 1,1,15,7,7,7,10,10,10,15)
ParamterMin = c(5, 50, 100, 150, 0.1, 1.3, 0.1,9, 0.1, 0.1, 0.1, 3,0.5, 0.5, 0.5 , 0.5,0.5,0.5,0.5)
ParamterModell = 0.5 * (ParamterMax + ParamterMin)
class(ParamterModell) <- "VIC"
VICPaList <- PaListMake(ParamterModell)
}
