% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ParamAll}
\alias{ParamAll}
\title{#' σ-Stefan-Boltzmann constant [4.903 10-9 MJ K-4 m-2 day-1]
"CONST_StefanBoltzmann"}
\format{
list of num
}
\usage{
ParamAll
}
\description{
#' Karman’s constant [0.40]
"CONST_Karman"
all parameters.
}
\keyword{datasets}
