% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation.R
\name{PRECDivid}
\alias{PRECDivid}
\title{Precipitation Divid}
\usage{
PRECDivid(InData, Param)
}
\arguments{
\item{InData}{indata list, use SNOWDivid(runMode = "VIEW") view the variables and theirs structures}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,
the other parameters depednd on the actuell model, eg. TimeStepSec, gridN.
and use SNOWDivid(runMode = "VIEW") view the structure}
}
\value{
use SNOWDivid(runMode = "VIEW") view the outputs and theirs structure
}
\description{
Determines from the air temperature what fraction of incoming
precipitation is frozen and unfrozen (snow and rain).
}
