% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUH.R
\name{PUH}
\alias{PUH}
\title{make period UH from IUH}
\usage{
PUH(f, lower, upper, ..., subdivisions = 100L)
}
\arguments{
\item{f}{list of parameter for every IUH}

\item{lower}{lower list of IUH methond name}

\item{upper}{upper boundry}

\item{...}{other parmters for IUH}

\item{subdivisions}{subdivisions}
}
\value{
list of Parameter for all UH mit class name
}
\description{
make period UH from IUH
}
