% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physic.R
\name{q_to_vp}
\alias{q_to_vp}
\title{q_to_vp}
\usage{
q_to_vp(q, p)
}
\arguments{
\item{q}{specific humidity}

\item{p}{pressure}
}
\value{
vp vapor pressure (units are the same as p)
}
\description{
convert specific humidity (q) to vapor pressure (vp) based on
         pressure (p)
}
