% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET.R
\name{ReferenceET.Hargreaves}
\alias{ReferenceET.Hargreaves}
\title{ReferenceET with Hargreaves methond}
\usage{
\method{ReferenceET}{Hargreaves}(InData, Param, runMode = "RUN", viewGN = 3, ...)
}
\arguments{
\item{InData}{list of: JDay, Latitude, Tmax, Tmin, Tmean
\itemize{
\item JDay, date
\item Latitude
\item Tmax, oC, Celius
\item Tmin, oC, Celius
\item Tmean, oC, Celius
}}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}

\item{...}{other Parmeters}
}
\value{
ReferenceET with Hargreaves methond
}
\description{
ReferenceET with Hargreaves methond
}
\references{
Hargreaves G H, Samani Z A. Reference crop evapotranspiration from ambient air temperature[J]. American Society of Agricultural Engineers, 1985(1):1-12.
}
