% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET.R
\name{ReferenceET.Linacre}
\alias{ReferenceET.Linacre}
\title{ReferenceET with PenmanMonteith methond}
\usage{
\method{ReferenceET}{Linacre}(InData, Param, ...)
}
\arguments{
\item{InData}{list of:
\itemize{
\item Elevation, m
\item Latitude
\item Tmean, oC, Celius
\item Actual_vapor_press,
}}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,}

\item{...}{other Parmeters}
}
\value{
Reference ET with PenmanMonteith methond
}
\description{
ReferenceET with PenmanMonteith methond
}
\references{
Linacre ET. 1977. A simple formula for estimating evaporation rates in various climates, using temperature data alone. AgriculturalMeteorology 18: 409-424.
}
