% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET.R
\name{ReferenceET.PenmanMonteith}
\alias{ReferenceET.PenmanMonteith}
\title{ReferenceET with PenmanMonteith methond}
\usage{
\method{ReferenceET}{PenmanMonteith}(InList, ...)
}
\arguments{
\item{InList}{list of:
\itemize{
\item JDay, date
\item Elevation, m
\item Latitude
\item Tmax, oC, Celius
\item Tmin, oC, Celius
\item Tmean, oC, Celius
\item Wind, speed, m/s
\item WindH, hoch, speed messeur
\item SunHour,
\item RelativeHumidity,
}}

\item{...}{other Paramater and inputdata}
}
\value{
Reference ET with PenmanMonteith methond
}
\description{
ReferenceET with PenmanMonteith methond
}
\references{
Pengman H L. Estimating evaporation[J]. American Geophysical Union, 1956(1):43-50.
}
