% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET.R
\name{ReferenceET.PmMt}
\alias{ReferenceET.PmMt}
\title{ReferenceET with PmMt methond}
\usage{
\method{ReferenceET}{PmMt}(InData, Param, runMode = "RUN", viewGN = 3, ...)
}
\arguments{
\item{InData}{list of:
\itemize{
\item JDay, date
\item Elevation, m
\item Latitude
\item TMax, oC, Celius
\item TMin, oC, Celius
\item TAir, oC, Celius
\item Wind, speed, m/s
\item WindH, hoch, speed messeur
\item SunHour,
\item RelativeHumidity,
}}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}

\item{...}{other Parmeters}
}
\value{
Reference ET with PmMt methond
}
\description{
ReferenceET with PmMt methond
}
\references{
Pengman H L. Estimating evaporation[J]. American Geophysical Union, 1956(1):43-50.
}
