% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET.R
\name{ReferenceET}
\alias{ReferenceET}
\title{ReferenceET caculate}
\usage{
ReferenceET(InList, ...)
}
\arguments{
\item{InList}{list of Input data}

\item{...}{other Paramater and inputdata}
}
\value{
ReferenceET
}
\description{
ReferenceET caculate
}
