% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUTE.R
\name{ROUTE.G2RES}
\alias{ROUTE.G2RES}
\title{use IUH calculate flow from grid to station}
\usage{
\method{ROUTE}{G2RES}(InData, Param, ...)
}
\arguments{
\item{InData}{indata list, use snow_density(runMode = "VIEW") view the variables and theirs structures
RunoffFlowList list of Inputdata(Runoff) [[1]] for surface flow, [[2]] for base flow
UHAll all of the UH
TypeGridID all of Grid type in geological matrix
TransAll all of translate matrix}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,}

\item{...}{other Parmeters}
}
\value{
station flow
}
\description{
use IUH calculate flow from grid to station
}
