% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUTE.R
\name{ROUTE.Gr4j}
\alias{ROUTE.Gr4j}
\title{route from GR4J}
\usage{
\method{ROUTE}{Gr4j}(InData, Param, ...)
}
\arguments{
\item{InData}{indata list, use snow_density(runMode = "VIEW") view the variables and theirs structures}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,}

\item{...}{other Parmeters}
}
\value{
station flow
}
\description{
route from GR4J
}
\references{
https://webgr.inrae.fr/en/models/daily-hydrological-model-gr4j/description-of-the-gr4j-model/

Perrin, C., Michel, C. and Andréassian, V., 2003. Improvement of a parsimonious model for streamflow simulation. Journal of Hydrology, 279 : 275-289, DOI: 10.1016/S0022-1694(03)00225-7
}
