% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUTE.R
\name{ROUTE}
\alias{ROUTE}
\title{from gridflow to station flow route with UH}
\usage{
ROUTE(InData, ...)
}
\arguments{
\item{InData}{indata list, use snow_density(runMode = "VIEW") view the variables and theirs structures}

\item{...}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,}
}
\description{
from gridflow to station flow route with UH
}
