% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUTE.R
\name{ROUTE.WSnGSmGn}
\alias{ROUTE.WSnGSmGn}
\title{caculate route with UH}
\usage{
\method{ROUTE}{WSnGSmGn}(InData, Param, ...)
}
\arguments{
\item{InData}{indata list, use snow_density(runMode = "VIEW") view the variables and theirs structures
RunoffFlowList list of Inputdata(Runoff)
UHAll all of the UH
TypeGridID all of Grid type in geological matrix
TransAll all of translate matrix
WSN    IUH the number of water source
GSN    IUH  the number of step from grid to station}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,}

\item{...}{other Parmeters}
}
\value{
StationFlow Q
}
\description{
caculate route with UH
}
