% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOWBlowing.R
\name{rtnewt}
\alias{rtnewt}
\title{rtnewt}
\usage{
rtnewt(x1, x2, acc, Ur, Zr, BLOWING_MAX_ITER)
}
\arguments{
\item{x1}{not klar}

\item{x2}{not klar}

\item{acc}{not klar}

\item{Ur}{not klar}

\item{Zr}{not klar}

\item{BLOWING_MAX_ITER}{not klar}
}
\value{
not klar
}
\description{
Newton-Raphson method.
the function modify from VIC original rescouce,
check the rescouce code in https://github.com/UW-Hydro/VIC
}
