% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{RUNOFF.Gr4j}
\alias{RUNOFF.Gr4j}
\title{VIC runoff (OverInfiltrationExcessRunoff and SaturationExcessRunoff)}
\usage{
\method{RUNOFF}{Gr4j}(InData, Param, ...)
}
\arguments{
\item{InData}{4-list of:
\itemize{
\item PrecipitationHoch
\item SoilMoistureVolume
\item AET
}}

\item{Param}{2-list of:
\itemize{
\item Gr4j_X1
}}

\item{...}{other Parmeters}
}
\value{
2-list of:
\itemize{
\item Runoff
\item MoistureVolume
}
}
\description{
VIC runoff (OverInfiltrationExcessRunoff and SaturationExcessRunoff)
}
\references{
https://webgr.inrae.fr/en/models/daily-hydrological-model-gr4j/description-of-the-gr4j-model/

Perrin, C., Michel, C. and Andréassian, V., 2003. Improvement of a parsimonious model for streamflow simulation. Journal of Hydrology, 279 : 275-289, DOI: 10.1016/S0022-1694(03)00225-7
}
