% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{RUNOFF}
\alias{RUNOFF}
\title{RUNOFF}
\usage{
RUNOFF(InList, ...)
}
\arguments{
\item{InList}{list of Input data}

\item{...}{other Paramater and inputdata}
}
\value{
SoilInfiltration
}
\description{
RUNOFF
}
