% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{RUNOFF.VIC}
\alias{RUNOFF.VIC}
\title{VIC runoff (OverInfiltrationExcessRunoff and SaturationExcessRunoff)}
\usage{
\method{RUNOFF}{VIC}(InList, PaList, ...)
}
\arguments{
\item{InList}{4-list of:
\itemize{
\item PrecipitationHoch
\item SoilMoistureCapacityMax
\item SoilMoistureVolume
\item InfiltrationRateMax
}}

\item{PaList}{2-list of:
\itemize{
\item paSoilMoistureCapacityB
\item paInfiltrationRateB
}}

\item{...}{other Paramater and inputdata}
}
\value{
2-list of:
\itemize{
\item Runoff
\item Infiltration
}
}
\description{
VIC runoff (OverInfiltrationExcessRunoff and SaturationExcessRunoff)
}
\references{
Liang X, Xie Z. A new surface runo parameterization with subgrid-scale soil heterogeneity for land surface models[J]. Advances in Water Resources, 2001(24):1173-1193.
}
