% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOWBlowing.R
\name{shear_stress}
\alias{shear_stress}
\title{shear_stress}
\usage{
shear_stress(U10, ZO, utshear, BLOWING_MAX_ITER)
}
\arguments{
\item{U10}{not klar}

\item{ZO}{not klar}

\item{utshear}{not klar}

\item{BLOWING_MAX_ITER}{not klar}
}
\value{
threshold shear stress
}
\description{
Iterate to find actual shear stress during saltation.
}
