% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOW.R
\name{SNOW.17}
\alias{SNOW.17}
\title{Snow model SNOW17}
\usage{
\method{SNOW}{`17`}(InData, Param, ...)
}
\arguments{
\item{InData}{2-list of:
\itemize{
\item TAir
\item RainFall
\item SnowFall
}}

\item{Param}{2-list of:
\itemize{
\item Factor_Day_degree
\item Base_T
}}

\item{...}{other Parmeters}
}
\value{
1-list of:
\itemize{
\item Precipitation
}
}
\description{
Snow model SNOW17
}
\references{
Anderson, E.A., 'National Weather Service River Forecast System - Snow Accumulation and Ablation Model', NOAA Technical Memorandum NWS HYDRO-17, 217 pp, November 1973.
}
