% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOW.R
\name{SNOW.Ddf}
\alias{SNOW.Ddf}
\title{Snow model day-degree-factor}
\usage{
\method{SNOW}{Ddf}(InData, Param, ...)
}
\arguments{
\item{InData}{2-list of:
\itemize{
\item TAir
\item RainFall
\item SnowFall
}}

\item{Param}{2-list of:
\itemize{
\item Factor_Day_degree
\item Base_T
}}

\item{...}{other Parmeters}
}
\value{
1-list of:
\itemize{
\item Precipitation
}
}
\description{
Snow model day-degree-factor
}
\references{
Hock, R. Temperature index melt modelling in mountain areas. J. Hydrol. 2003, 282, 104–115.
}
