% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOW.R
\name{SNOW}
\alias{SNOW}
\title{SNOW}
\usage{
SNOW(InData, ...)
}
\arguments{
\item{InData}{indata list, use SNOWIntercept(runMode = "VIEW") view the variables and theirs structures}

\item{...}{other arguments}
}
\value{
use SNOW(runMode = "VIEW") view the outputs and theirs structure
}
\description{
This routine was written to handle the various calls and data
handling needed to solve the various components of the new VIC
snow code for both the full_energy and water_balance models.
}
