% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOW.R
\name{SNOW.VIC}
\alias{SNOW.VIC}
\title{SNOW}
\usage{
\method{SNOW}{VIC}(InData, Param, runMode = "RUN", viewGN = 3, ...)
}
\arguments{
\item{InData}{indata list, use SNOW(runMode = "VIEW") view the variables and theirs structures
param Options Optionslist, alle the options for model is TURE seted as default,
and use SNOW(runMode = "VIEW") view the options structure and set the options}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,
the other parameters depednd on the actuell model, eg. TimeStepSec, gridN.
and use SNOW(runMode = "VIEW") view the structure}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}

\item{...}{other Parmeters}
}
\value{
use SNOW(runMode = "VIEW") view the outputs and theirs structure
}
\description{
This routine was written to handle the various calls and data
handling needed to solve the various components of the new VIC
snow code for both the full_energy and water_balance models.
}
