% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOWOther.R
\name{snow_albedo}
\alias{snow_albedo}
\title{snow_albedo}
\usage{
snow_albedo(
  dt,
  new_snow,
  swq,
  cold_content,
  last_snow,
  MELTING,
  Param,
  runMode = "RUN"
)
}
\arguments{
\item{dt}{temstep in seconds}

\item{new_snow}{new snow}

\item{swq}{snow water}

\item{cold_content}{cold content}

\item{last_snow}{last snow}

\item{MELTING}{logical, ob melting}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,
the other parameters depednd on the actuell model, eg. TimeStepSec, gridN.
and use snow_albedo(runMode = "VIEW") view the structure}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}
}
\value{
use snow_albedo(runMode = "VIEW") view the outputs and theirs structure
}
\description{
This subroutine computes the snow pack surface albedo.'
}
\references{
Computes albedo as a function of snow age and season, based on the
algorithm of the US Army Corps of Engineers.
}
