% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOWOther.R
\name{snow_density}
\alias{snow_density}
\title{snow_density}
\usage{
snow_density(InData, Param, runMode = "RUN", viewGN = 3)
}
\arguments{
\item{InData}{indata list, use snow_density(runMode = "VIEW") view the variables and theirs structures}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,
the other parameters depednd on the actuell model, eg. TimeStepSec, gridN.
and use snow_density(runMode = "VIEW") view the structure}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}
}
\value{
use snow_density(runMode = "VIEW") view the outputs and theirs structure
}
\description{
Compute the snow density based on swe and snow metamorphism.
}
\references{
DENS_SNTHRM <- Algorithm is taken from SNTHERM89, adjusted for an essentially single-layer model.
}
