% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOWBlowing.R
\name{SNOWBlowing}
\alias{SNOWBlowing}
\title{SNOWBlowing}
\usage{
SNOWBlowing(InData, Param, runMode = "RUN", viewGN = 3)
}
\arguments{
\item{InData}{indata list, use SNOWIntercept(runMode = "VIEW") view the variables and theirs structures}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,
the other parameters depednd on the actuell model, eg. TimeStepSec, gridN.
and use SNOWIntercept(runMode = "VIEW") view the structure
and use SNOWBlowing(runMode = "VIEW") view the options structure and set the options}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}
}
\value{
use SNOWIntercept(runMode = "VIEW") view the outputs and theirs structure
}
\description{
Calculate energy of sublimation from blowing snow.
Calculate sublimation from blowing snow
}
