% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOWIntercept.R
\name{SNOWIntercept}
\alias{SNOWIntercept}
\title{SNOWIntercept}
\usage{
SNOWIntercept(InData, Param, runMode = "RUN", viewGN = 3)
}
\arguments{
\item{InData}{indata list, use SNOWIntercept(runMode = "VIEW") view the variables and theirs structures}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,
the other parameters depednd on the actuell model, eg. TimeStepSec, gridN.
and use SNOWIntercept(runMode = "VIEW") view the structure}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}
}
\value{
use SNOWIntercept(runMode = "VIEW") view the outputs and theirs structure
}
\description{
Calculates the interception and subsequent release of by the forest canopy
using an energy balance approach.
}
\references{
Determine the maximum snow interception water equivalent.
Kobayashi, D., 1986, Snow Accumulation on a Narrow Board,
Cold Regions Science and Technology, (13), pp. 239-245. Figure 4.
}
