% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnergyBalance.R
\name{SnowPackEnergyBalance}
\alias{SnowPackEnergyBalance}
\title{SnowPackEnergyBalance}
\usage{
SnowPackEnergyBalance(InData, Param)
}
\arguments{
\item{InData}{list of Input data}

\item{Param}{Paramaters and inputdata}
}
\value{
a out list
}
\description{
Calculate the surface energy balance for the snow pack.
}
