% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnergyBalance.R
\name{StabilityCorrection}
\alias{StabilityCorrection}
\title{StabilityCorrection}
\usage{
StabilityCorrection(Z, d, TSurf, Tair, Wind, Z0)
}
\arguments{
\item{Z}{ref_height}

\item{d}{displacement}

\item{TSurf}{tempratur of snow surface}

\item{Tair}{tempratur of air}

\item{Wind}{wind speed}

\item{Z0}{roughness}
}
\value{
Correction
}
\description{
Calculate atmospheric stability correction for non-neutral conditions
}
