% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOWBlowing.R
\name{sub_with_height}
\alias{sub_with_height}
\title{sub_with_height}
\usage{
sub_with_height(z, es, Wind, EactAir, FT, hsalt, phi_r, ushear, Param)
}
\arguments{
\item{z}{not klar}

\item{es}{not klar}

\item{Wind}{not klar}

\item{EactAir}{not klar}

\item{FT}{not klar}

\item{hsalt}{not klar}

\item{phi_r}{not klar}

\item{ushear}{not klar}

\item{Param}{not klar}
}
\value{
the sublimation rate for a given height above the boundary layer
}
\description{
Calculate the sublimation rate for a given height above the boundary layer.
}
