% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOWBlowing.R
\name{transport_with_height}
\alias{transport_with_height}
\title{transport_with_height}
\usage{
transport_with_height(z, Wind, ZO, hsalt, phi_r, ushear, Param)
}
\arguments{
\item{z}{not klar}

\item{Wind}{not klar}

\item{ZO}{not klar}

\item{hsalt}{not klar}

\item{phi_r}{not klar}

\item{ushear}{not klar}

\item{Param}{not klar}
}
\value{
transport rate for a given height above the boundary layer
}
\description{
Calculate the transport rate for a given height above the boundary layer.
}
