% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASEFLOW.R
\name{BASEFLOW.ARNO}
\alias{BASEFLOW.ARNO}
\title{baseflow}
\usage{
\method{BASEFLOW}{ARNO}(InList, PaList, ...)
}
\arguments{
\item{InList}{2-list of:
\itemize{
\item SoilMoistureVolume
\item SoilMoistureVolumeMax
}}

\item{PaList}{4-list of:
\itemize{
\item paDrainageLossMax
\item paDrainageLossMin
\item paExponentARNOBase
\item paSoilMoistureVolumeARNOBaseThresholdRadio
}}

\item{...}{other Paramater and inputdata}
}
\value{
baseflow
}
\description{
baseflow
}
\references{
Arnold J G, Srinivasan R, Muttiah R S et al. Large area hydrologic modeling and assessmentpart I: Model Develoment [J]. Journal of the American Water Resources Association, 1998(34):73-89.
}
