% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASEFLOW.R
\name{BASEFLOW}
\alias{BASEFLOW}
\title{caculate Base flow}
\usage{
BASEFLOW(InList, ...)
}
\arguments{
\item{InList}{list of Input data}

\item{...}{other Paramater and inputdata}
}
\value{
BASEFLOW
}
\description{
caculate Base flow
}
