% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUTE.R
\name{CONFLUENCE.G2RES}
\alias{CONFLUENCE.G2RES}
\title{caculate route with UH}
\usage{
\method{CONFLUENCE}{G2RES}(InData, Param, ...)
}
\arguments{
\item{InData}{indata list, use snow_density(runMode = "VIEW") view the variables and theirs structures
RunoffFlowList list of Inputdata(Runoff)
UHAll all of the UH
TypeGridID all of Grid type in geological matrix
TransAll all of translate matrix}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,}

\item{...}{other Parmeters}
}
\value{
StationFlow Q
}
\description{
caculate route with UH
}
