% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUTE.R
\name{CONFLUENCE}
\alias{CONFLUENCE}
\title{caculate route with UH}
\usage{
CONFLUENCE(InData, ...)
}
\arguments{
\item{InData}{indata list, use SNOWIntercept(runMode = "VIEW") view the variables and theirs structures}

\item{...}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,}
}
\value{
StationFlow Q
}
\description{
caculate route with UH
}
