% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET.R
\name{ET.VIC}
\alias{ET.VIC}
\title{Actual evapotranspiration in VIC}
\usage{
\method{ET}{VIC}(InData, Param, runMode = "RUN", viewGN = 3, ...)
}
\arguments{
\item{InData}{9-list of:
\itemize{
\item ReferenceEvap,
\item PrecipitationHoch,
\item MoistureVolume,
\item CapacityMax,
\item Volume1,
\item CapacityMax1,
\item AerodynamicResistance,
\item ArchitecturalResistance,
\item StomatalResistance,
}}

\item{Param}{1-list of:
\itemize{
\item paramSoilMoistureCapacityB
}}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}

\item{...}{other Parmeters}
}
\value{
Actual evapotranspiration
}
\description{
Actual evapotranspiration in VIC
}
\references{
Liang X, Lettenmaier D P. A simple hydrologically based model of land surface water and energy fluxes for general circulation models[J]. Journal of Geophysical Research, 1994(99):415-428.
}
