% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GROUNDWATER.R
\name{GROUNDWATER.VIC}
\alias{GROUNDWATER.VIC}
\title{ground water VIC function}
\usage{
\method{GROUNDWATER}{VIC}(InList, PaList, ...)
}
\arguments{
\item{InList}{6-list of:
\itemize{
\item Evapotranspiration
\item Interception
\item Infiltration
\item BaseFlow
\item GroundWaterIn
\item GridSoilParame
}}

\item{PaList}{1-list of:
\itemize{
\item paClappHornbergerB
}}

\item{...}{other Paramater and inputdata}
}
\value{
INTERCEPTION
}
\description{
ground water VIC function
}
\references{
Liang X, Xie Z, Huang M. A new parameterization for surface and groundwater interactions and its impact on water budgets with the variable infiltration capacity[J]. Journal of Geochemical Research, 2003(108):1-17.
}
