% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{Infiltration.OIER}
\alias{Infiltration.OIER}
\title{Infiltration rate in Over Infiltration Excess Runoff}
\usage{
\method{Infiltration}{OIER}(InData, Param, runMode = "RUN", viewGN = 3, ...)
}
\arguments{
\item{InData}{2-list of:
\itemize{
\item PrecipitationHoch
\item InfiltrationRateMax
}}

\item{Param}{1-list of:
\itemize{
\item paInfiltrationRateB
}}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}

\item{...}{other Parmeters}
}
\value{
Infiltration rate in Over Infil tration Excess Runoff
}
\description{
Infiltration rate in Over Infiltration Excess Runoff
}
