% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{Infiltration.SER}
\alias{Infiltration.SER}
\title{SoilInfiltration rate in Saturation Excess Runoff modell}
\usage{
\method{Infiltration}{SER}(InData, Param, ...)
}
\arguments{
\item{InData}{3-list of:
\itemize{
\item PrecipitationHoch
\item SoilMoistureCapacityMax
\item SoilMoistureCapacity
}}

\item{Param}{1-list of:
\itemize{
\item paSoilMoistureCapacityB
}}

\item{...}{other Parmeters}
}
\value{
SoilInfiltration in SaturationExcessRunoff modell
}
\description{
SoilInfiltration rate in Saturation Excess Runoff modell
}
