% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{InfiltratRat.GreenAmpt}
\alias{InfiltratRat.GreenAmpt}
\title{Infiltration with GreenAmpt methond}
\usage{
\method{InfiltratRat}{GreenAmpt}(InData, Param, runMode = "RUN", viewGN = 3, ...)
}
\arguments{
\item{InData}{5-list of:
\itemize{
\item HydraulicConductivity
\item WettingFrontSoilSuction
\item SoilMoistureContent
\item EffectivePorosity
\item SoilMoistureVolume
}}

\item{Param}{Param}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}

\item{...}{other Parmeters}
}
\value{
Infiltration rat
}
\description{
Infiltration with GreenAmpt methond
}
\references{
Tarboton D G. Rainfall-runoff processes[M]. Utah State University, 2003.
}
