% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERCEPTION.R
\name{INTERCEPTION.Gash}
\alias{INTERCEPTION.Gash}
\title{INTERCEPTION whith Gash methond}
\usage{
\method{INTERCEPTION}{Gash}(InData, Param, runMode = "RUN", viewGN = 3, ...)
}
\arguments{
\item{InData}{4-list of:
\itemize{
\item Volum,
\item CanopyStorageCapacity,
\item RainfallDuringSaturation,
\item Evaporation
}}

\item{Param}{1-list of:
\itemize{
\item paCoefficientFreeThroughfall
}}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}

\item{...}{other Parmeters}
}
\value{
INTERCEPTION
}
\description{
INTERCEPTION whith Gash methond
}
\references{
Gash J H C. An analytical model of rainfall interception by forests[J]. Quarterly Journal of the Royal Meteorological Society, 1978(105):43-55.
}
