% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERCEPTION.R
\name{INTERCEPTION}
\alias{INTERCEPTION}
\title{INTERCEPTION caculate}
\usage{
INTERCEPTION(InList, ...)
}
\arguments{
\item{InList}{list of Input data}

\item{...}{other Paramater and inputdata}
}
\value{
INTERCEPTION
}
\description{
INTERCEPTION caculate
}
