% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitation.R
\name{PRECDivid}
\alias{PRECDivid}
\title{Precipitation Divid}
\usage{
PRECDivid(InData, Param, runMode = "RUN", viewGN = 3)
}
\arguments{
\item{InData}{indata list, use SNOWDivid(runMode = "VIEW") view the variables and theirs structures}

\item{Param}{paramlist, in this R packege ParamAll dataset there are alredy most parameters,
the other parameters depednd on the actuell model, eg. TimeStepSec, gridN.
and use SNOWDivid(runMode = "VIEW") view the structure}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}
}
\value{
use SNOWDivid(runMode = "VIEW") view the outputs and theirs structure
}
\description{
Determines from the air temperature what fraction of incoming
precipitation is frozen and unfrozen (snow and rain).
}
