% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET.R
\name{ReferenceET.Hargreaves}
\alias{ReferenceET.Hargreaves}
\title{ReferenceET with Hargreaves methond}
\usage{
\method{ReferenceET}{Hargreaves}(InList, ...)
}
\arguments{
\item{InList}{list of: JDay, Latitude, Tmax, Tmin, Tmean
\itemize{
\item JDay, date
\item Latitude
\item Tmax, oC, Celius
\item Tmin, oC, Celius
\item Tmean, oC, Celius
}}

\item{...}{other Paramater and inputdata}
}
\value{
ReferenceET with Hargreaves methond
}
\description{
ReferenceET with Hargreaves methond
}
\references{
Hargreaves G H, Samani Z A. Reference crop evapotranspiration from ambient air temperature[J]. American Society of Agricultural Engineers, 1985(1):1-12.
}
