% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{RUNOFF.OIER}
\alias{RUNOFF.OIER}
\title{Infiltration in OverInfiltrationExcessRunoff}
\usage{
\method{RUNOFF}{OIER}(InList, PaList, ...)
}
\arguments{
\item{InList}{3-list of:
\itemize{
\item PrecipitationHoch
\item SoilMoistureCapacityMax
\item SoilMoistureVolum
}}

\item{PaList}{1-list of:
\itemize{
\item paSoilMoistureCapacityB
}}

\item{...}{other Paramater and inputdata}
}
\value{
2-list of:
\itemize{
\item Runoff
\item Infiltration
}
}
\description{
Infiltration in OverInfiltrationExcessRunoff
}
