% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{RUNOFF.VIC}
\alias{RUNOFF.VIC}
\title{VIC runoff (OverInfiltrationExcessRunoff and SaturationExcessRunoff)}
\usage{
\method{RUNOFF}{VIC}(InData, Param, runMode = "RUN", viewGN = 3, ...)
}
\arguments{
\item{InData}{4-list of:
\itemize{
\item PrecipitationHoch
\item SoilMoistureCapacityMax
\item SoilMoistureVolume
\item InfiltrationRateMax
}}

\item{Param}{2-list of:
\itemize{
\item paSoilMoistureCapacityB
\item paInfiltrationRateB
}}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}

\item{...}{other Parmeters}
}
\value{
2-list of:
\itemize{
\item Runoff
\item Infiltration
}
}
\description{
VIC runoff (OverInfiltrationExcessRunoff and SaturationExcessRunoff)
}
\references{
Liang X, Xie Z. A new surface runo parameterization with subgrid-scale soil heterogeneity for land surface models[J]. Advances in Water Resources, 2001(24):1173-1193.
}
