% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{RUNOFF.VM}
\alias{RUNOFF.VM}
\title{VitcalMisingExcessRunoff runoff (OverInfiltrationExcessRunoff and SaturationExcessRunoff)}
\usage{
\method{RUNOFF}{VM}(InData, Param, ...)
}
\arguments{
\item{InData}{4-list of:
\itemize{
\item PrecipitationHoch
\item SoilMoistureCapacityMax
\item SoilMoistureCapacity
\item InfiltrationRateMax
}}

\item{Param}{2-list of:
\itemize{
\item paSoilMoistureCapacityB
\item paInfiltrationRateB
}}

\item{...}{other Parmeters}
}
\value{
2-list of:
\itemize{
\item Runoff
\item Infiltration in VitcalMisingExcessRunoff
}
}
\description{
VitcalMisingExcessRunoff runoff (OverInfiltrationExcessRunoff and SaturationExcessRunoff)
}
