% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNOFF.R
\name{RUNOFF.VM}
\alias{RUNOFF.VM}
\title{VitcalMisingExcessRunoff runoff (OverInfiltrationExcessRunoff and SaturationExcessRunoff)}
\usage{
\method{RUNOFF}{VM}(InData, Param, runMode = "RUN", viewGN = 3, ...)
}
\arguments{
\item{InData}{4-list of:
\itemize{
\item PrecipitationHoch
\item SoilMoistureCapacityMax
\item SoilMoistureCapacity
\item InfiltrationRateMax
}}

\item{Param}{2-list of:
\itemize{
\item paSoilMoistureCapacityB
\item paInfiltrationRateB
}}

\item{runMode}{mode to run the function, there three mode:
\itemize{
\item "RUN": default, run the function like general faunction
\item "VIEW": view the structures of Arguments and Output(return)
\item "CHECK": chek the structure of the Arguments
}}

\item{viewGN}{grid nummer for "VIEW" mode.}

\item{...}{other Parmeters}
}
\value{
2-list of:
\itemize{
\item Runoff
\item Infiltration in VitcalMisingExcessRunoff
}
}
\description{
VitcalMisingExcessRunoff runoff (OverInfiltrationExcessRunoff and SaturationExcessRunoff)
}
